# main.py
# Import necessary modules and libraries
import os
import openai
from oandapyV20 import API
from oandapyV20.exceptions import V20Error
from oandapyV20.endpoints.instruments import InstrumentsCandles
from flask import Flask, jsonify, request, send_from_directory
from flask_cors import CORS, cross_origin
from flask import make_response
import numpy as np
from datetime import datetime
from datetime import timedelta

# Setup Flask app
app = Flask(__name__)

# Enable CORS and configure CORS Headers
cors = CORS(app, resources={
    r"/*": {
        "origins": "https://chat.openai.com",  # Update this to the origin you want to allow
        "allow_headers": [
            "Content-Type",
            "Authorization",
            "Access-Control-Allow-Credentials",
            "Access-Control-Allow-Headers",
            "Access-Control-Allow-Methods",
            "Access-Control-Allow-Origin",
            "Baggage",
            "sentry-trace",
            "openai-conversation-id",
            "openai-ephemeral-user-id"  
        ]
    }
})

app.config['CORS_HEADERS'] = 'Content-Type'

# Routes for different functionalities of the application
@app.route('/generate', methods=['POST'])
# Generate endpoint uses OpenAI API to generate text based on a given prompt and temperature
def generate():
    data = request.get_json()
    prompt = data.get('prompt')
    temperature = data.get('temperature', 0.5)
    generated_text = openai.Completion.create(engine="text-davinci-003", 
    prompt=prompt, 
    temperature=temperature,
    max_tokens=4096
    )
    return jsonify({'generated_text': generated_text})

@app.route('/complete', methods=['POST'])
# Complete endpoint uses OpenAI API to complete given text 
def complete():
    data = request.get_json()
    text = data.get('text')
    completed_text = openai.Completion.create(model="text-davinci-003",
    text=text,
    max_tokens=4096
    )
    return jsonify({'completed_text': completed_text})

@app.route('/search', methods=['POST'])
# Search endpoint uses OpenAI API to search based on a given query
def search():
    data = request.get_json()
    query = data.get('query')
    response = openai.Completion.create(
        engine="gpt-3.5-turbo-0301",
        prompt=query,
        max_tokens=4096
    )
    generated_text = response.choices[0].text.strip()
    return jsonify({'results': generated_text})

@app.route('/playground', methods=['POST'])
# Playground endpoint uses OpenAI API to generate code based on a given prompt
def playground():
    data = request.get_json()
    code = data.get('code')
    response = openai.Completion.create(
        engine="gpt-3.5-turbo-0301",
        prompt=code,
        max_tokens=4096
    )
    generated_text = response.choices[0].text.strip()
    return jsonify({'output': generated_text})

@app.route('/logo.png')
# Serves logo image
def serve_logo():
    return send_from_directory('.', 'logo.png', mimetype='image/png')

@app.route('/openapi.yaml')
# Serves OpenAPI specification
def serve_openai_yaml():
    return send_from_directory('.', 'openapi.yaml', mimetype='text/yaml')

@app.route('/.well-known/ai-plugin.json')
# Serves AI plugin manifest
def serve_ai_plugin_manifest():
    return send_from_directory('.well-known', 'ai-plugin.json', mimetype='application/json')

@app.route('/prices', methods=['GET', 'POST'])
# Prices endpoint fetches forex price data from OANDA API
def get_prices():
    # Handle both GET and POST requests
    if request.method == 'POST':
        data = request.get_json()
    else:  # It's a GET request
        data = request.args.to_dict()
    # ... Remaining code for this function ...

# Class for interfacing with OpenAI API and OANDA API
class OpenAIPlugin(object):
    def __init__(self, oanda_api_key, openai_api_key):
        # Initialize with API keys and setup API clients
        self.oanda_api_key = oanda_api_key
        self.openai_api_key = openai_api_key
        self.oanda_client = API(access_token=self.oanda_api_key, environment="practice")
        openai.api_key = self.openai_api_key

    # ... Remaining methods for this class ...

# Start the Flask application
if __name__ == "__main__":
    app.run(host='0.0.0.0', port=5003)